renaming, organizing, correcting color on font, fixed text, changed color on clinic boxes, fixed spelling, hover and more
