+++ package.json
@@ -13,7 +13,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@mui/material": "^5.14.19",
+    "@mui/material": "^5.14.20",
     "@mui/styled-engine-sc": "^6.0.0-alpha.7",
     "framer-motion": "^10.16.10",
     "lottie-react": "^2.4.0",
