+++ src/App.tsx
@@ -1,19 +1,19 @@
 import React, { useState } from "react";
 import { Burger, Menu, Navbar, Logo } from "./components/navbar";
 import "./index.css";
-import { ContentWrapper, BackgroundColor } from "./components/ContentWrapper";
+import { ContentWrapper } from "./components/ContentWrapper";
 import Frontpage from "./components/frontpage";
 import cloud from "./assets/cloudmain.svg";
 import { CloudContainer, Cloud } from "./components/frontpage/cloud";
-import Payment from "./components/paymentSite/payment";
-import NoIdPage from "./components/signIn/NoId";
-import Ticket from "./components/Ticket/Ticket";
-import useTicket from "./components/Ticket/useTicket";
+import Payment from "./components/payment/payment";
+import IdPage from "./components/signIn/NoId";
+import Ticket from "./components/ticket/Ticket";
+import useTicket from "./components/ticket/useTicket";
 import About from "./components/About";
-import Contact from "./components/Contact";
-import ChooseClinic from "./components/chooseClinic/ChooseClinic";
+import Contact from "./components/contact/Contact";
+import ChooseClinic from "./components/clinic/ChooseClinic";
 import TextWrapper from "./components/TextWrapper";
-import hospital from "./assets/WalkInClinic.png"
+
 
 function App() {
   const [open, setOpen] = React.useState(false);
@@ -30,45 +30,25 @@ function App() {
   };
 
   const pages: { [page: string]: JSX.Element } = {
-    clinic: <ChooseClinic setNav={setNav} setClinic={setClinic} />, //'name of page' and component
-    ticket: (
-      <Ticket
-        ticketNumber={ticketNumber}
-        lastNumber={lastNumber}
-        waitTime={waitTime.toString()} //this has to be changed so it can pass as a string
-        requestNewTicket={requestNewTicket}
-        setNav={setNav}
-        clinic={clinic}
-      />
-    ),
-    about: <About />,
-    contact: <Contact setNav={setNav} />,
-    "sign-in": <NoIdPage setNav={setNav} nav={nav} />,
-    "sign-in-two": <NoIdPage setNav={setNav} nav={nav} />,
-    frontpage: <Frontpage setNav={setNav} />,
-    payment: <Payment setNav={setNav} clinic={clinic} />,
+    "clinic": <ChooseClinic setNav={setNav} setClinic={setClinic} />, //'name of page' and component.  
+    "ticket": (<Ticket ticketNumber={ticketNumber} lastNumber={lastNumber} 
+            waitTime={waitTime.toString()} //this has to be changed so it can pass as a string
+            requestNewTicket={requestNewTicket} setNav={setNav} clinic={clinic}/>),
+    "about": <About />,
+    "contact": <Contact setNav={setNav} />,
+    "sign-in": <IdPage setNav={setNav} nav={nav} />,
+    "sign-in-two": <IdPage setNav={setNav} nav={nav} />,
+    "frontpage": <Frontpage setNav={setNav} />,
+    "payment": <Payment setNav={setNav} clinic={clinic} />,
   };
 
   return (
     <>
-    
        <CloudContainer>
-        <Cloud
-          alt="fluffy white clouds moving in the background"
-          initial={{ x: -100 }}
-          animate={{ x: 0 }}
-          transition={{
-            repeat: Infinity,
-            repeatType: "mirror",
-            duration: 20,
-            ease: [0.67, 0.67, 0.83, 0.67],
-          }}
-          src={cloud}
-        />
+        <Cloud alt="fluffy white clouds moving in the background" initial={{ x: -100 }} animate={{ x: 0 }} 
+        transition={{repeat: Infinity, repeatType: "mirror", duration: 20, ease: [0.67, 0.67, 0.83, 0.67],}} src={cloud}/>
       </CloudContainer>
-     
       <ContentWrapper>
-       
         <Navbar>
           <Logo setNav={navChange} />
           <Burger open={open} setOpen={setOpen} />
@@ -77,9 +57,7 @@ function App() {
         <TextWrapper>
           {pages[nav]} {/*Navigation function for site */}
         </TextWrapper>
-        
       </ContentWrapper>
-      
     </>
   );
 }
