+++ src/components/About.tsx
@@ -1,3 +1,5 @@
+import '../index.css'
+
 
 const About = () => {
     return (
@@ -11,7 +13,7 @@ const About = () => {
             Instead of sitting in a room full of sick individuals, you can wait in a designated area—safe, socially distanced, and stress-free.</p>
             </div>
             <h1>OUR TEAM</h1>
-            <div style={{padding:"0 5"}}>
+            <div style={{padding:"0 5%"}}>
             <p>Pétur Jack, Sólveig Alda Halldórsdóttir and Marcela Slobodaniuc are not just tech enthusiasts; we're passionate about improving healthcare. <br />
             The design magic makes our website intuitive and delightful to use. We are all about user experience.
             We invite you to be part of this healthcare revolution. <br />Together, let's eliminate waiting times, prioritize your well-being, and create a healthier world.</p>
