+++ src/components/Button.tsx
@@ -22,7 +22,7 @@ import styled from "styled-components";
 `
 
 const AppLoginButton = styled(ButtonPrimary)`
-    color: black;
+    color: #293437;
     background-color:white;
     font-size:1.6rem;
     height: 45px;
