+++ src/components/TextWrapper.tsx
@@ -1,5 +1,5 @@
 
-// TextWrapper is made as a style for all pages with text elements in between components. 
+// TextWrapper is made as a style for all pages that have text elements in between components.  It works on About page, contact page and ticket page. 
 
 import React from "react";
 import styled from "styled-components"
@@ -10,8 +10,12 @@ const StyledTextWrapper = styled.div`
     justify-content: center;
     gap: 2rem;
     align-items: center;
+
     @media (max-width: 768px) {
-    align-items: center;
+        align-items: center;
+        p {
+            max-width: 300px;
+        }
     }
 
 `
