+++ src/components/clinic/BoxWrapper.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const BoxWrapper = styled.div`
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 1rem;
+`
